Fixed label positioning
When the segment only contained 2 points, the text label was drawn at
the first segment point. As suggested after testing, the middle point of
these 2 is now computed and the label is drawn there.
Still, for the cases where the segment has more than one point, the
middle point from the segment array is selected. This point does not
always coincide with the middle point of the segment, but because the
segment can be bent, the middle point will not be computed.